src/main/webapp/app/entities/abc-14/delete/abc-14-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Abc14Service } from '../service/abc-14.service';

import { Abc14DeleteDialogComponent } from './abc-14-delete-dialog.component';

describe('Abc14 Management Delete Component', () => {
  let comp: Abc14DeleteDialogComponent;
  let fixture: ComponentFixture<Abc14DeleteDialogComponent>;
  let service: Abc14Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [Abc14DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Abc14DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Abc14DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Abc14Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
